+++ src/OrdersList/hooks/useBuildQuery/useBuildQuery.js
@@ -45,6 +45,7 @@ export function useBuildQuery(customFields) {
     'sortby metadata.updatedDate/sort.descending',
     {
       [FILTERS.DATE_CREATED]: buildDateTimeRangeQuery.bind(null, [FILTERS.DATE_CREATED]),
+      [FILTERS.DATE_UPDATED]: buildDateTimeRangeQuery.bind(null, [FILTERS.DATE_UPDATED]),
       [FILTERS.RENEWAL_DATE]: buildDateRangeQuery.bind(null, [FILTERS.RENEWAL_DATE]),
       [FILTERS.DATE_ORDERED]: buildDateTimeRangeQuery.bind(null, [FILTERS.DATE_ORDERED]),
       [FILTERS.CLOSE_REASON]: (filterValue) => {