+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change history for ui-orders
 
+* Add additional order filters to support reporting requirements. Refs UIOR-
+
 ## [6.0.2](https://github.com/folio-org/ui-orders/tree/v6.0.2) (2024-04-01)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v6.0.1...v6.0.2)